+++ projects/komponentkartan/src/lib/controls/list-item/list-item.component.ts
@@ -1,6 +1,6 @@
 import {
   Component, Input, Output, EventEmitter, ContentChildren, ContentChild, QueryList,
-  AfterContentInit, forwardRef, OnDestroy, OnChanges, SimpleChanges,
+  AfterContentInit, forwardRef, OnDestroy, OnChanges, SimpleChanges, ElementRef, Renderer,
 } from '@angular/core';
 import { trigger, style, transition, animate, state } from '@angular/animations';
 
@@ -97,7 +97,7 @@ export class ListItemComponent implements AfterContentInit, OnDestroy, OnChanges
 
   @ContentChildren(forwardRef(() => ListColumnComponent), { descendants: true }) columns: QueryList<ListColumnComponent>;
 
-  constructor(private listService: ListService) {
+  constructor(private listService: ListService, private elementRef: ElementRef, private renderer: Renderer) {
     this.expandedChanged.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.hideNotifications());
   }
 
@@ -119,13 +119,6 @@ export class ListItemComponent implements AfterContentInit, OnDestroy, OnChanges
   }
 
   ngAfterContentInit() {
-    if (this.listItemHeader) {
-      this.listItemHeader.expandedChanged.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.toggleExpanded());
-      this.listItemHeader.goToFirst.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnFirstRow.emit());
-      this.listItemHeader.goToLast.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnLastRow.emit());
-      this.listItemHeader.goUp.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnPreviousRow.emit());
-      this.listItemHeader.goDown.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnNextRow.emit());
-    }
     if (this.listContent) {
       this.listContent.goUp.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnPreviousRowContent.emit());
       this.listContent.goDown.pipe(takeUntil(this.ngUnsubscribe)).subscribe(() => this.setFocusOnNextRowContent.emit());
@@ -144,22 +137,47 @@ export class ListItemComponent implements AfterContentInit, OnDestroy, OnChanges
       this.listService.requestExpandListItem(this);
     }
   }
+  keydown(event: KeyboardEvent) {
+    if (event.key === 'Enter' || event.key === 'Spacebar' || event.key === ' ') {
+      this.toggleExpanded();
+      event.preventDefault();
+    }
+    if (event.key === 'Home') {
+      this.setFocusOnFirstRow.emit();
+      event.preventDefault();
+    }
+    if (event.key === 'End') {
+      this.setFocusOnLastRow.emit();
+      event.preventDefault();
+    }
+    if ((event.ctrlKey && event.key === 'PageUp') || event.key === 'ArrowUp' || event.key === 'Up') {
+      this.setFocusOnPreviousRow.emit();
+      event.preventDefault();
+    }
+    if ((event.ctrlKey && event.key === 'PageDown') || event.key === 'ArrowDown' || event.key === 'Down') {
+      this.setFocusOnNextRow.emit();
+      event.preventDefault();
+    }
+  }
 
   setExpanded(expanded: boolean) {
     if (!this.notInteractable) {
       this.isExpanded = expanded;
       this.expandedChanged.emit(this.isExpanded);
+
       this.notInteractable = true;
       setTimeout(() => { this.notInteractable = false; }, 400);
     }
   }
 
   setFocusOnRow() {
-    this.listItemHeader.setFocus();
+      const item = this.elementRef.nativeElement.querySelector('.list-item__header_wrapper');
+      item.focus();
   }
 
   hideNotifications() {
     if (this.temporaryNotificationVisible) {
+      this.setFocusOnRow();
       if (this.notification && this.notification.type === NotificationType.ShowOnRemove) {
         setTimeout(() => {
           this.isDeleted = true;
@@ -184,19 +202,20 @@ export class ListItemComponent implements AfterContentInit, OnDestroy, OnChanges
     }
   }
 
+  /** Hantering av färg på vänsterkanten. Sätts till success/error om valt, annars fallback */
   handleNotificationColor() {
     const current = this.temporaryNotification ? this.temporaryNotification : this.permanentNotification;
-    if (current) {
-      // Hantera färg på vänsterkanten
-      if (current.icon === 'vgr-icon-exclamation--red' || current.icon === 'vgr-icon-ok-check-green') {
-        this.notificationColor = current.icon === 'vgr-icon-exclamation--red' ? 'notification-error' : 'notification-success';
-      } else {
-        this.notificationColor = null;
-      }
+    if (current && ('icon' in current)) {
+      if (current.icon.color === 'success') {
+        this.notificationColor = 'notification-success';
+      } else if (current.icon.color === 'error') {
+        this.notificationColor = 'notification-error';
+      } else { this.notificationColor = null; }
     } else {
       this.notificationColor = null;
     }
   }
+
 
   triggerDeletedEvent() {
     if (this.isDeleted) {
@@ -224,5 +243,4 @@ export class ListItemComponent implements AfterContentInit, OnDestroy, OnChanges
     this.ngUnsubscribe.next();
     this.ngUnsubscribe.complete();
   }
-
 }