+++ projects/komponentkartan/src/lib/models/rowNotification.model.ts
@@ -1,6 +1,9 @@
 import { NotificationType } from './notificationType.model';
+import { NotificationIconTypes } from './notificationIconTypes';
+
+
 export interface RowNotification {
-    icon: string;
+    icon: NotificationIconTypes;
     type: NotificationType;
     message: string;
     done: boolean;