+++ src/app/lists/examples/examples.ts
@@ -223,21 +223,22 @@ export class Examples {
 
       this.peopleRows = this.examplePeople.map(x => new ExpandableRow<ExamplePerson, ExamplePerson>(x));
 
-      this.peopleRows[0].setNotification('Meddelande: Text', 'vgr-icon-message');
-      this.peopleRows[4].setNotification('Personen är inaktiv', 'vgr-icon-exclamation--red');
-    }
+      this.peopleRows[0].setNotification('Meddelande: Text', { name: 'comment-dots' } as NotificationIconTypes);
+      this.peopleRows[2].setNotification('tillfällig', { name: 'comment-dots' } as NotificationIconTypes, true);
+      this.peopleRows[4].setNotification('Personen är inaktiv', { name: 'exclamation-circle', color: 'error', solid: true } as NotificationIconTypes);
+      }
 
     deleteRow(row: ExpandableRow<ExamplePerson, ExamplePerson>) {
       // Remove visually.
-      row.notifyOnRemove(row.previewObject.firstName + ' togs bort och kommer inte längre att kunna logga in', 'vgr-icon-ok-check');
+      row.notifyOnRemove(row.previewObject.firstName + ' togs bort och kommer inte längre att kunna logga in', { name: 'check-circle', color: 'success' });
     }
 
     updateRow(row: ExpandableRow<ExamplePerson, ExamplePerson>) {
-      row.notifyOnCollapse(row.previewObject.firstName + ' sparades', 'vgr-icon-ok-check-green');
+      row.notifyOnCollapse(row.previewObject.firstName + ' sparades', { name: 'check-circle', color: 'success' });
     }
 
     updateRow2(row: ExpandableRow<ExamplePerson, ExamplePerson>) {
-      row.notifyOnCollapse(row.previewObject.firstName + ' sparades', 'vgr-icon-ok-check-green', true);
+      row.notifyOnCollapse(row.previewObject.firstName + ' sparades', { name: 'check-circle', color: 'success' }, true);
     }
 
     onSortChanged(event: SortChangedArgs) {
@@ -285,9 +286,8 @@ export class Examples {
       <vgr-list-column width="5">{{row.firstName}}</vgr-list-column>
       <vgr-list-column width="5">{{row.lastName}}</vgr-list-column>
       <vgr-list-column width="5" align="right">{{row.income | number:'2.2-2':'sv-SE'}}</vgr-list-column>
-      <vgr-list-column-trashcan [disabled]="row.previewObject.deleted" (delete)="onDeleteRow(row)" width="1"></vgr-list-column-trashcan>
-      <vgr-list-column-checkbox [disabled]="row.previewObject.deleted" [checked]="row.previewObject.selected" (checkedChanged)="onSelectRowChanged(row, $event)"
-        width="3"></vgr-list-column-checkbox>
+      <vgr-list-column-trashcan [disabled]="row.previewObject.deleted" (click)="onDeleteRow($event,row)" (keydown)="handleKeyDown($event,row)" [align]="'center'"></vgr-list-column-trashcan>
+      <vgr-list-column-checkbox width="3" [disabled]="row.previewObject.deleted" [checked]="row.previewObject.selected" (checkedChanged)="onSelectRowChanged(row, $event)"></vgr-list-column-checkbox>  width="3"></vgr-list-column-checkbox>
     </vgr-list-item-header>
     <vgr-list-item-content>
       <span>Mer information</span>
@@ -363,8 +363,14 @@ export class Examples {
             }
         }
 
-        onDeleteRow(row: any) {
-            this.removeRow(row);
+        onDeleteRow(event: Event, row: any) {
+          event.stopPropagation();
+          this.removeRow(row);
+        }
+        handleKeyDown(event: KeyboardEvent, row: any) {
+          if (event.key === 'Enter' || event.key === 'Spacebar' || event.key === ' ') {
+              this.onDeleteRow(event, row);
+          }
         }
 
         notifyOnDelete(row: any) {
@@ -379,7 +385,7 @@ export class Examples {
         }
 
         removeSelectedRow() {
-            this.rowToRemove.notifyOnRemove(this.rowToRemove.previewObject.firstName + ' togs bort', 'vgr-icon-ok-check');
+            this.rowToRemove.notifyOnRemove(this.rowToRemove.previewObject.firstName + ' togs bort', , { name: 'check-circle', color: 'success' });
             this.rowToRemove.previewObject.selected = false;
             this.rowToRemove.previewObject.deleted = true;
             /*
@@ -472,7 +478,11 @@ export class Examples {
 
         this.listNotification = {
           message: 'Här är ett exempel på en list-notifikation. De kan användas om det t.ex. blir något fel när man hämtar datan från servicen.',
-          icon: 'vgr-icon-exclamation--red'
+          icon: {
+            name: 'exclamation-circle',
+            color: 'error',
+            solid: true
+          }
         };
 
       }