+++ src/app/lists/examples/withactionbuttons/listexamplewithactionbuttons.component.ts
@@ -56,6 +56,11 @@ export class ListExampleWithActionButtonsComponent {
         event.stopPropagation();
         this.removeRow(row);
     }
+    handleKeyDown(event: KeyboardEvent, row: any) {
+        if (event.key === 'Enter' || event.key === 'Spacebar' || event.key === ' ') {
+            this.onDeleteRow(event, row);
+        }
+    }
 
     notifyOnDelete(row: any) {
         this.removedObjectString = JSON.stringify(row, null, '\t');
@@ -77,7 +82,7 @@ export class ListExampleWithActionButtonsComponent {
     }
 
     removeSelectedRow() {
-        this.rowToRemove.notifyOnRemove(this.rowToRemove.previewObject.firstName + ' togs bort', 'vgr-icon-ok-check');
+        this.rowToRemove.notifyOnRemove(this.rowToRemove.previewObject.firstName + ' togs bort', { name: 'check-circle', color: 'success' });
         this.rowToRemove.previewObject.selected = false;
         this.rowToRemove.previewObject.deleted = true;
         /*